+++ app/concepts/matestack/ui/core/app/app.js
@@ -16,15 +16,15 @@ const componentDef = {
   }),
   mounted: function(){
     const self = this;
-    window.onpopstate = (event) => {
+    window.addEventListener("popstate", (event) => {
       if (isNavigatingToAnotherPage({
           origin: self.currentOrigin,
           pathName: self.currentPathName,
           search: self.currentSearch
         }, document.location)){
-        self.$store.dispatch("navigateTo", {url: document.location.pathname, backwards: true} );
+        self.$store.dispatch("navigateTo", { url: document.location.pathname + document.location.search, backwards: true } );
       }
-    }
+    })
   },
   components: {
     VRuntimeTemplate: VRuntimeTemplate