+++ spec/dummy/app/models/test_model.rb
@@ -1,2 +1,5 @@
 class TestModel < ApplicationRecord
+
+  has_one_attached :file
+
 end