+++ spec/usage/base/app_spec.rb
@@ -158,6 +158,60 @@ describe "App", type: :feature, js: true do
 
   end
 
+  it "has access to view context" do
+
+    class Apps::ExampleApp < Matestack::Ui::App
+
+      def response
+        components {
+          heading size: 1, text: "My Example App Layout"
+          if @view_context.view_renderer.instance_of?(ActionView::Renderer)
+            plain "has access to ActionView Context"
+          end
+          plain link_to "Test Link", "/some/page" # calling an ActionView Url Helper here
+          plain time_ago_in_words(3.minutes.from_now) # calling an ActionView Date Helper here
+          main do
+            page_content
+          end
+        }
+      end
+
+    end
+
+    module Pages::ExampleApp
+
+    end
+
+    class Pages::ExampleApp::ExamplePage < Matestack::Ui::Page
+
+      def response
+        components {
+          div id: "my-div-on-page-1" do
+            heading size: 2, text: "This is Page 1"
+          end
+        }
+      end
+
+    end
+
+
+    class ExampleAppPagesController < ExampleController
+      include Matestack::Ui::Core::ApplicationHelper
+
+      def page1
+        responder_for(Pages::ExampleApp::ExamplePage)
+      end
+
+    end
+
+    visit "app_specs/my_example_app/page1"
+
+    expect(page).to have_content("has access to ActionView Context")
+    expect(page).to have_content("Test Link")
+    expect(page).to have_content("3 minutes")
+
+  end
+
   it "can navigate back using browser history"
 
   it "just uses serverside routes, which works standalone"